fix(ci): lan doctest paths + macOS u32→c_ulong + main.rs doc indent
Fourth CI round on PR #13 — 3 residual failures (Deno still red, out
of scope):

- Test: brainwires-network/src/lan/mod.rs doctest used `network::*` as
  its module qualifier, a leftover from when the LAN tooling lived in
  brainwires-hardware::network. Inside the doctest `use brainwires_
  network::lan;` exposes the type at `lan::*`. Replaced 4 instances
  (`network::list_interfaces`, `network::get_ip_configs`,
  `network::scan_common_ports`, `network::PortState`).
- macOS: brainwires-cli/src/session/server.rs:146 — libc::TIOCSCTTY
  is u32 on macOS/BSD but u64 on Linux. `libc::ioctl` expects
  `c_ulong` (always u64). Cast: `libc::TIOCSCTTY as libc::c_ulong`.
- Clippy: brainwires-chat-pwa/home/src/main.rs:6-7 — doc list overin-
  dented (3+ spaces for continuation lines). Reformatted the
  `* pair` bullet's continuation to 2-space indent.

Verified locally:
- RUSTFLAGS=-Dwarnings cargo clippy --workspace -- -D warnings exit 0
- RUSTFLAGS=-Dwarnings cargo test --workspace --doc exit 0

+++ extras/brainwires-cli/src/session/server.rs
@@ -143,7 +143,7 @@ impl SessionServer {
 
                 // Set controlling terminal
                 unsafe {
-                    libc::ioctl(0, libc::TIOCSCTTY, 0);
+                    libc::ioctl(0, libc::TIOCSCTTY as libc::c_ulong, 0);
                 }
 
                 // Exec the TUI
